意见簿回复功能，所有人都可以回复 意见簿回复内容删除，只有本人和管理员可以删除 院长信息回复内容删除调整

+++ dms/src/main/java/com/whr/dms/service/SuggestionReplyServiceImpl.java
@@ -2,19 +2,27 @@ package com.whr.dms.service;
 
 import java.util.List;
 
-import javax.annotation.Resource;
-
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.whr.dms.dao.TReplyDao;
+import com.whr.dms.dao.TSuggestionDao;
+import com.whr.dms.exceptions.ParameterCheckException;
+import com.whr.dms.models.SuggestionType;
 import com.whr.dms.models.TReply;
+import com.whr.dms.models.TSuggestion;
+import com.whr.dms.security.RoleType;
+import com.whr.dms.security.SecurityUtil;
 
 @Service
 public class SuggestionReplyServiceImpl implements SuggestionReplyService {
 
-	@Resource
+	@Autowired
 	private TReplyDao rdao;
+	@Autowired
+	private TSuggestionDao sdao;
 	
 	@Override
 	@Transactional(readOnly = true)
@@ -36,14 +44,51 @@ public class SuggestionReplyServiceImpl implements SuggestionReplyService {
 
 	@Override
 	@Transactional
-	public void deleteSuggestionReply(long replyId) {
-		rdao.delete(replyId);
+	public void deleteSuggestionReply(long replyId) throws ParameterCheckException {
+		TReply r = rdao.findOne(replyId);
+		
+		if (r == null) {
+			throw new ParameterCheckException("未找到记录");
+		}
+		
+		TSuggestion s = sdao.findOne(r.getSuggestionId());
+		// 是否是作者本人
+		boolean isMe = SecurityUtil.isMe(r.getAuthorId());
+
+		// 是否为业务管理员
+		boolean isManager = SecurityUtil.hasRole(getNeededRole(s.getType()));
+
+		if (isMe || isManager) {
+			// 有操作权
+			rdao.delete(replyId);
+		} else {
+			// 拒绝
+			throw new AccessDeniedException("没有操作权限。");
+		}
+		
 	}
 
 	@Override
 	@Transactional
 	public List<TReply> getSuggestionReply(long suggestionId) {
 		return rdao.getReplyList(suggestionId);
+	}
+	
+	/**
+	 * 根据意见簿类型（意见簿、院长信箱、医院管理）查找需要的角色
+	 * 
+	 * @param type
+	 * @return
+	 */
+	private RoleType getNeededRole(SuggestionType type) {
+		if (SuggestionType.Suggestion.equals(type)) {
+			return RoleType.ROLE_SUGGESTION_MANAGER;
+		} else if (SuggestionType.President.equals(type)) {
+			return RoleType.ROLE_PRESIDENT_MANAGER;
+		} else if (SuggestionType.Managment.equals(type)) {
+			return RoleType.ROLE_HOSPITAL_MANAGER;
+		}
+		throw new RuntimeException("意见簿类型匹配错误");
 	}
 
 }
